Fix fact test and add fact testing for history
A test for one fact requires that the previous two results not be
complex.  This commit adds that requirement.

While fixing the above bug, the fact testing module could not
handle checking more than the most recent calculation.  It was
modified so that a history of calculations can be included.