Transition disco music correctly
Previously, there were two problems with the disco music.  One was
that the audio would sometimes get stuck on the last moments of
the song.  The other was that the same song would repeat instead
of selecting a new song.  These were caused by incorrectly
detecting when the end of the song was reached.  This commit
introduces a variable to track what the state of the song was, to
determine whether to continue playing the currently-selected song
or to pick a new song.